Add explicit versionend events
This allows us to dispatch the finishdownload event before downloading
the next patch. This, in turn, means that if there is a flaky network
connection and we can't download the next patch we can still commit the
currently downloaded data. It also means, we can reset the retry timeout
so that the subsequent download has a shorter rety timeout.