fix: Convert download stream to an async generator instead
This should hopefully fix some of the hangs we were seeing when we'd get
overlapping database updates and downloads in Firefox as well as making
the behavior more understandable--notably that we don't start
downloading version N+1 until we have finished applying version N.

This is because when using streams, the browser will try to fill the
buffer and because the download stream is just one long stream of all
the versions, as soon as we've finished downloading one version, we'll
start to fill the buffer with the next version.

Ideally we'd download and write at the same time but because IndexedDB
doesn't allow transactions to be held open indefinitely we have to do
the download and update steps independently so we can update a version
as a single transaction.

This change also makes the code quite a bit simpler.