feat: Allow matching names on kana equivalence
e.g. allow searching for まるた to match the record for マルタ.

Technically we should make the upgrade procedure here generate the
hiragana keys for all the existing records. However, we haven't released
the names version of this library widely yet and this feature is mostly
only needed for Rikaichamp (where this hasn't been released yet).

Also, upgrading all 740k records would take a _lot_ of time and we don't
yet have a means of reporting upgrade status so for now we don't bother.

The 1~2 users on version 2 of the database can just clobber the database
if they really need this feature.